Checks is username exists when signing up

+++ server/routes/signup.js
@@ -8,36 +8,48 @@ module.exports = function(req, res) {
     fs.readFile("./data/users.json", "utf8", function(err, data){
         let existingJSON = JSON.parse(data);
 
-        let userID = existingJSON.length + 1;
-        
-        let newUser = {
-            "id": userID,
-            "username": username,
-            "password": password,
-            "email": email,
-            "roles": ["US"],
-            "groups": ["open"]
-        };
-    
-        existingJSON.push(newUser);
-    
-        let jsonString = JSON.stringify(existingJSON);
-
-        fs.writeFile("./data/users.json", jsonString, err => {
-            if (err){
-                console.log(err);
-                res.send({successful: false});
-            } else {
-                let response = {
-                    valid: true,
-                    username: newUser.username,
-                    email: newUser.email,
-                    id: newUser.id,
-                    groups: newUser.groups,
-                    roles: newUser.roles
-                };
-                res.send(response);
+        // Check username exists
+        let existing = false;
+        for (let user of existingJSON){
+            if (user.username == username){
+                existing = true;
+                break;
             }
-        });
+        }
+
+        if (existing == true){
+            res.send({valid: false});
+        } else {
+            let userID = existingJSON.length + 1;
+        
+            let newUser = {
+                "id": userID,
+                "username": username,
+                "password": password,
+                "email": email,
+                "roles": ["US"],
+                "groups": ["open"]
+            };
+        
+            existingJSON.push(newUser);
+        
+            let jsonString = JSON.stringify(existingJSON);
+    
+            fs.writeFile("./data/users.json", jsonString, err => {
+                if (err){
+                    res.send({successful: false});
+                } else {
+                    let response = {
+                        valid: true,
+                        username: newUser.username,
+                        email: newUser.email,
+                        id: newUser.id,
+                        groups: newUser.groups,
+                        roles: newUser.roles
+                    };
+                    res.send(response);
+                }
+            });
+        }
     });
 }
